requirements.txt
phidata==24.3.1
yfinance==0.2.51
duckduckgo-search==7.2.1